+++ xtask/src/ci.rs
@@ -69,6 +69,7 @@ enum FeatureSet {
     Markdown,
     Socks,
     SsoLogin,
+    ImageProc,
 }
 
 #[derive(Subcommand, PartialEq, Eq, PartialOrd, Ord)]
@@ -218,6 +219,7 @@ fn run_feature_tests(cmd: Option<FeatureSet>) -> Result<()> {
         (FeatureSet::Markdown, "--features markdown,testing"),
         (FeatureSet::Socks, "--features socks,testing"),
         (FeatureSet::SsoLogin, "--features sso-login,testing"),
+        (FeatureSet::ImageProc, "--features image-proc,testing"),
     ]);
 
     let run = |arg_set: &str| {