+++ crates/matrix-sdk/CHANGELOG.md
@@ -11,6 +11,10 @@ Breaking changes:
   `Media::get_file`/`Media::remove_file`/`Media::get_thumbnail`/`Media::remove_thumbnail`
 - A custom sliding sync proxy set with `ClientBuilder::sliding_sync_proxy` now takes precedence over a discovered proxy.
 - `Client::get_profile` was moved to `Account` and renamed to `Account::fetch_user_profile_of`. `Account::get_profile` was renamed to `Account::fetch_user_profile`.
+- `generate_image_thumbnail` now returns a `Thumbnail`.
+- It is now possible to select the format of a generated thumbnail.
+  - `generate_image_thumbnail` takes a `ThumbnailFormat`.
+  - `AttachmentConfig::generate_thumbnail` takes a `ThumbnailFormat`.
 
 Additions: